pyproject.toml
[project]
name = "crewavaliadora"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.157.0",
    "crewai-tools>=0.60.0",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
]
